supportmax-pro/v2-cognitive/pyproject.toml
[project]
name = "supportmax-pro-v2-cognitive"
version = "2.0.0"
description = "Cognitive Enhancement Implementation"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.109.0",
    "uvicorn==0.27.0",
    "pydantic==2.6.0",
    "pydantic-settings==2.1.0",
    "openai>=1.13.3",
    "anthropic==0.18.0",
    "python-dotenv==1.0.0",
    "requests==2.31.0",
    "pytest==8.0.0",
    "httpx==0.26.0",
    "streamlit==1.31.0",
    "crewai==0.19.0",
    "langchain-openai==0.0.5",
    "chromadb>=0.4.22",
    "sentence-transformers>=2.3.1",
    "langchain-community>=0.0.19",
    "pypdf>=4.0.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
